ecommerce_api/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
drf-spectacular
pillow